refactor: landing layout refactor.
