+++ src/assets/images/index.ts
@@ -1,11 +1,10 @@
 import auth_bg from "./auth-bg.png";
 import heroImage from './hero-heading.png';
 import companyLogo from './Nomad.png';
-import desktop from './desktop.png';
 import socialMedia from './social-media.png';
 import CompanyLogo from "./company-logo.jpg";
 import CompanyPic1 from "./company-pic1.jpg";
 import CompanyPic2 from "./company-pic2.jpg";
 import CompanyPic3 from "./company-pic3.jpg";
 
-export { CompanyLogo, CompanyPic1, CompanyPic2, CompanyPic3, heroImage, companyLogo, desktop, socialMedia, auth_bg  };
+export { CompanyLogo, CompanyPic1, CompanyPic2, CompanyPic3, heroImage, companyLogo, socialMedia, auth_bg  };
