+++ src/components/layouts/landing/index.ts
@@ -1,3 +1,3 @@
-export * from './header'
+export * from './Header'
 export * from './main'
-export * from './footer'
+export * from './Footer'
