+++ src/components/sections/landing/hero.tsx
@@ -1,55 +1,48 @@
-import { MapPin, Search } from "lucide-react";
-import { useState } from "react";
+import { LocalIcon } from "@/assets/icons/local-icon";
+import { LocalImage } from "@/assets/images/local-image";
 
 export const Hero = () => {
-  const [searchQuery, setSearchQuery] = useState("");
-  const [location, setLocation] = useState("");
+
   return (
-    <section className="bg-gradient-to-br from-blue-50 to-indigo-50 py-20">
-      <div className="max-w-7xl mx-auto px-6">
-        <div className="text-center mb-12">
-          <h1 className="text-5xl font-bold text-gray-900 mb-4">
-            Discover more than <span className="text-blue-600">5000+ Jobs</span>
-          </h1>
-          <p className="text-xl text-gray-600 mb-8 max-w-2xl mx-auto">
-            Great platform for the job seeker that are passionate about
-            startups. Find your dream job easier.
-          </p>
-
-          {/* Search Bar */}
-          <div className="bg-white rounded-lg shadow-lg p-2 max-w-4xl mx-auto mb-8">
-            <div className="flex flex-col md:flex-row gap-2">
-              <div className="flex-1 flex items-center px-4 py-3 border-r border-gray-200">
-                <Search className="w-5 h-5 text-gray-400 mr-3" />
-                <input
-                  type="text"
-                  placeholder="Job title or keyword"
-                  value={searchQuery}
-                  onChange={(e) => setSearchQuery(e.target.value)}
-                  className="flex-1 outline-none text-gray-700"
-                />
-              </div>
-              <div className="flex-1 flex items-center px-4 py-3">
-                <MapPin className="w-5 h-5 text-gray-400 mr-3" />
-                <input
-                  type="text"
-                  placeholder="Florence, Italy"
-                  value={location}
-                  onChange={(e) => setLocation(e.target.value)}
-                  className="flex-1 outline-none text-gray-700"
-                />
-              </div>
-              <button className="bg-blue-600 hover:bg-blue-700 text-white px-8 py-3 rounded-lg font-medium transition-colors">
-                Search my job
-              </button>
-            </div>
+      <section className="flex flex-col items-center justify-center gap-4 py-8 px-[124px] w-full max-w-[2000px]">
+        <LocalImage
+          imageName="heroImage"
+          className="w-[450px] h-[70px] md:w-[850px] md:h-[130px] object-cover"
+        />
+        <h1 className="text-[#515B6F] text-lg font-normal ">
+          Find the dream companies you dream work for
+        </h1>
+        <div className="flex flex-col md:flex-row gap-4 justify-between w-full max-w-[950px] shadow-xl rounded p-6 bg-white">
+          <div className="flex items-center gap-2">
+            <LocalIcon
+              iconName="search"
+              width={24}
+              height={24}
+              className="left-4 top-3 text-gray-500"
+            />
+            <input
+              type="text"
+              placeholder="Company name"
+              className="w-full px-4 py-2 border-b border-[#D6DDEB] rounded focus:outline-none focus:ring-0"
+            />
           </div>
-
-          <p className="text-gray-500 text-sm">
-            Popular: UI Designer, UX Researcher, Android, Admin
-          </p>
+          <div className="flex items-center gap-2">
+            <LocalIcon
+              iconName="Location"
+              width={24}
+              height={24}
+            />
+            <select className="w-full px-4 py-2 border-b border-[#D6DDEB] rounded text-gray-700">
+              <option value="">Select location</option>
+              <option value="florence">Florence, Italy</option>
+              <option value="paris">Paris, France</option>
+              <option value="newyork">New York, USA</option>
+            </select>
+          </div>
+          <button className="bg-[#4640DE] text-white px-4 py-2 rounded cursor-pointer">
+            Search
+          </button>
         </div>
-      </div>
-    </section>
+      </section>
   );
 };
